Improve syntax highlighting, better options
Assembled info files don't have the `^_` character anymore, so we have
to match the header line of the next node instead. For simplicity the
regex is stored once and then the syntax commands are assembled as
strings.

The info-specific options have been set up to match the indentation and
alignment rules used by Info.