Move private details out of public namespace
Some of the commands are now no longer defined in the 'plugin/info.vim'
file because that forced me to make functions public that don't belong
into the public namespace. It also allowed for some more cleanup of the
code.

In particular, we now have a duality between URIs and node references.