Allow :Info without arguments
Info without arguments acts like standalone info. In the process of
getting this to work I had to remove a number of mechanics that protect
the plugin from its author's stupidity. The plugin is no more naive and
will not try to fill in the missing file or node when building a
reference.