Fix URI not opening manual
This issue was that the autocommand was set up in the lazily loaded
portion of the code. I have added an autocommand to the initial loading
which will lazily load the script when an Info URI is opened.

The initial autocommands are now grouped and will be removed after the
script has been fully loaded. This will clean up the autocommands when
they are no longer needed.